refactor: make eip-4844 mandatory in backend
